Try catch and validation added

+++ CM.bll/Services/ContactGroupService.cs
@@ -21,10 +21,16 @@ namespace CM.bll.Services
 
         public void Save(ContactGroup entity)
         {
+
+
             try
             {
+                if (entity is null) throw new ArgumentNullException(nameof(entity));
+
                 entity.Active = true;
                 entity.CreatedDate = DateTime.Now;
+
+                ApplyValidation(entity);
                 _repo.ContactGroupRepo.Add(entity);
                 _repo.Save();
             }
@@ -33,49 +39,110 @@ namespace CM.bll.Services
                 Console.WriteLine(e);
                 throw;
             }
-           
+
         }
 
         public void Update(ContactGroup entity)
         {
-            var existingEntity = _repo.ContactGroupRepo.GetById(entity.Id);
-            if (existingEntity != null)
+            try
             {
-                existingEntity.Name = entity.Name;
-           
-                existingEntity.ModifiedDate = DateTime.Now;
-                _repo.ContactGroupRepo.Update(existingEntity);
-                _repo.Save();
+                if (entity is null) throw new ArgumentNullException(nameof(entity));
+                var existingEntity = _repo.ContactGroupRepo.GetById(entity.Id);
+
+                if (existingEntity != null)
+                {
+                    existingEntity.Name = entity.Name;
+
+                    existingEntity.ModifiedDate = DateTime.Now;
+                    ApplyValidation(existingEntity);
+                    _repo.ContactGroupRepo.Update(existingEntity);
+                    _repo.Save();
+                }
             }
+            catch (Exception)
+            {
+
+                throw;
+            }
+
         }
 
         public void DeleteById(long id)
         {
-            var entity = _repo.ContactGroupRepo.GetById(id);
-            _repo.ContactGroupRepo.Remove(entity);
-            _repo.Save();
+            try
+            {
+                if (id <= 0) throw new Exception("Id Should greater than zero");
+
+                var entity = _repo.ContactGroupRepo.GetById(id);
+                _repo.ContactGroupRepo.Remove(entity);
+                _repo.Save();
+            }
+            catch (Exception)
+            {
+
+                throw;
+            }
+
         }
 
         public ContactGroup FindById(long id)
         {
-            return _repo.ContactGroupRepo.GetById(id);
+            try
+            {
+                if (id <= 0) throw new Exception("Id Should greater than zero");
+                return _repo.ContactGroupRepo.GetById(id);
+            }
+            catch (Exception)
+            {
+
+                throw;
+            }
+
         }
 
         public IEnumerable<ContactGroup> Get()
         {
-           return _repo.ContactGroupRepo.GetAll();
+            try
+            {
+                return _repo.ContactGroupRepo.GetAll();
+            }
+            catch (Exception)
+            {
+
+                throw;
+            }
+
         }
 
 
         public IEnumerable<ContactGroup> GetFilterable(ContactGroupFilterModel filter)
         {
-            return _repo.ContactGroupRepo.GetFilterable(filter);
+
+            try
+            {
+                if (filter is null) throw new ArgumentNullException(nameof(filter));
+
+                return _repo.ContactGroupRepo.GetFilterable(filter);
+            }
+            catch (Exception)
+            {
+
+                throw;
+            }
+
+        }
+        private void ApplyValidation(ContactGroup entity)
+        {
+            if (entity is null) throw new ArgumentNullException(nameof(entity));
+
+            if (string.IsNullOrEmpty(entity.Name)) throw new Exception("Name is required");
+            
         }
         public override void Dispose()
         {
             _repo.Dispose();
         }
 
-       
+
     }
 }
